[FEATURE] #57 : WappProfileCard
